Added tabs in vulnerability view
